fix: Align to esbonio startup changes
In the upcoming release of the esbonio language server there are some
breaking changes in the way the server is started. This commit fixes the
construction of the Language Client so that the server is started
correctly.

A number of cli arguments have been removed and the server now requires
a number of `initializationOptions` at startup

- `esbonio.sphinx.confDir`
- `esbonio.sphinx.srcDir`
- `esbonio.sphinx.buildDir` (previously was `--cache-dir`)
- `esbonio.server.logLevel` (previously was `--log-level`)
- `esbonio.server.logFilter` (previously was `--log-filter`)
- `esbonio.server.hideSphinxOutput` (previously was `--hide-sphinx-output`)

The version of the language server in `package.json` has also been
updated to the next version number.